Move video-data.ts
