+++ src/data/video-data.ts
@@ -1,30 +1,30 @@
-import BloodVesselCase1MP4 from "./tissue/heart/BloodVesselCase1.mp4";
-import BloodVesselCase2MP4 from "./tissue/heart/BloodVesselCase2.mp4";
-import BloodVesselCase3MP4 from "./tissue/heart/BloodVesselCase3.mp4";
-import BloodVesselCase4MP4 from "./tissue/heart/BloodVesselCase4.mp4";
-import BloodCellCase1MP4 from "./cell/heart/BloodCellCase1Time1.mp4";
-import BloodCellCase2MP4 from "./cell/heart/BloodCellCase1Time2.mp4";
-import BloodCellCase3MP4 from "./cell/heart/BloodCellCase1Time3.mp4";
-import BloodCellCase4MP4 from "./cell/heart/BloodCellCase1Time4.mp4";
+import BloodVesselCase1MP4 from "../assets/videos/tissue/heart/BloodVesselCase1.mp4";
+import BloodVesselCase2MP4 from "../assets/videos/tissue/heart/BloodVesselCase2.mp4";
+import BloodVesselCase3MP4 from "../assets/videos/tissue/heart/BloodVesselCase3.mp4";
+import BloodVesselCase4MP4 from "../assets/videos/tissue/heart/BloodVesselCase4.mp4";
+import BloodCellCase1MP4 from "../assets/videos/cell/heart/BloodCellCase1Time1.mp4";
+import BloodCellCase2MP4 from "../assets/videos/cell/heart/BloodCellCase1Time2.mp4";
+import BloodCellCase3MP4 from "../assets/videos/cell/heart/BloodCellCase1Time3.mp4";
+import BloodCellCase4MP4 from "../assets/videos/cell/heart/BloodCellCase1Time4.mp4";
 
-import NeuronsCase2 from "./tissue/brain/NeuronsCase2.mp4";
-import NeuronsCase3 from "./tissue/brain/NeuronsCase3.mp4";
-import SynapseCase1 from "./cell/brain/SynapseCase1Faster.mp4";
-import SynapseCase2 from "./cell/brain/SynapseCase2.mp4";
-import SynapseCase3 from "./cell/brain/SynapseCase3.mp4";
-import SynapseCase4 from "./cell/brain/SynapseCase4.mp4";
+import NeuronsCase2 from "../assets/videos/tissue/brain/NeuronsCase2.mp4";
+import NeuronsCase3 from "../assets/videos/tissue/brain/NeuronsCase3.mp4";
+import SynapseCase1 from "../assets/videos/cell/brain/SynapseCase1Faster.mp4";
+import SynapseCase2 from "../assets/videos/cell/brain/SynapseCase2.mp4";
+import SynapseCase3 from "../assets/videos/cell/brain/SynapseCase3.mp4";
+import SynapseCase4 from "../assets/videos/cell/brain/SynapseCase4.mp4";
 
-import NoseCase1 from "./tissue/nose/NoseCase1.mp4";
-import NoseCase2 from "./tissue/nose/NoseCase2.mp4";
-import NoseCase3 from "./tissue/nose/NoseCase3.mp4";
-import NoseCase4 from "./tissue/nose/NoseCase4.mp4";
-import NoseCellCase11 from "./cell/nose/NoseCellCase1Time1.mp4";
-import NoseCellCase12 from "./cell/nose/NoseCellCase1Time2.mp4";
-import NoseCellCase13 from "./cell/nose/NoseCellCase1Time3.mp4";
-import NoseCellCase14 from "./cell/nose/NoseCellCase1Time4.mp4";
-import NoseCellCase21 from "./cell/nose/NoseCellCase2Time1.mp4";
-import NoseCellCase22 from "./cell/nose/NoseCellCase2Time2.mp4";
-import NoseCellCase23 from "./cell/nose/NoseCellCase2Time3.mp4";
+import NoseCase1 from "../assets/videos/tissue/nose/NoseCase1.mp4";
+import NoseCase2 from "../assets/videos/tissue/nose/NoseCase2.mp4";
+import NoseCase3 from "../assets/videos/tissue/nose/NoseCase3.mp4";
+import NoseCase4 from "../assets/videos/tissue/nose/NoseCase4.mp4";
+import NoseCellCase11 from "../assets/videos/cell/nose/NoseCellCase1Time1.mp4";
+import NoseCellCase12 from "../assets/videos/cell/nose/NoseCellCase1Time2.mp4";
+import NoseCellCase13 from "../assets/videos/cell/nose/NoseCellCase1Time3.mp4";
+import NoseCellCase14 from "../assets/videos/cell/nose/NoseCellCase1Time4.mp4";
+import NoseCellCase21 from "../assets/videos/cell/nose/NoseCellCase2Time1.mp4";
+import NoseCellCase22 from "../assets/videos/cell/nose/NoseCellCase2Time2.mp4";
+import NoseCellCase23 from "../assets/videos/cell/nose/NoseCellCase2Time3.mp4";
 
 const HeartCellCase1 = {
   0: BloodCellCase1MP4,
